Move API client JS files to their own subfolder
